format_dt

Turns a data frame or matrix into a DT formatted table with export buttons for copy, csv or excel.

Usage

format_dt(x)

Arguments

x

a data object (either a matrix or a data.frame)

Value

a data table decorated with export buttons

Author

Denis O'Meally

Examples

format_dt(USArrests)